Refactor into translator folder
