+++ frontend/src/components/translator/translation-list/TranslationList.js
